Picks to ch_type. Methods for epochs and evoked.
