+++ mne/viz/raw.py
@@ -13,18 +13,16 @@ from functools import partial
 import numpy as np
 
 from ..externals.six import string_types
-from ..io.pick import pick_types, _pick_data_channels, pick_info, channel_type
+from ..io.pick import pick_types, _pick_data_channels, pick_info
 from ..io.proj import setup_proj
 from ..utils import verbose, get_config
 from ..time_frequency import psd_welch
 from .topo import _plot_topo, _plot_timeseries, _plot_timeseries_unified
-from .topomap import _check_outlines
 from .utils import (_toggle_options, _toggle_proj, tight_layout,
                     _layout_figure, _plot_raw_onkey, figure_nobar,
                     _plot_raw_onscroll, _mouse_click, plt_show,
                     _helper_raw_resize, _select_bads, _onclick_help)
 from ..defaults import _handle_default
-from ..channels.layout import _auto_topomap_coords
 
 
 def _plot_update_raw_proj(params, bools):
@@ -815,94 +813,4 @@ def plot_raw_psd_topo(raw, tmin=0., tmax=None, fmin=0, fmax=100, proj=False,
         plt_show(show, block=block)
     except TypeError:  # not all versions have this
         plt_show(show)
-    return fig
-
-
-def plot_sensors(info, kind='topomap', picks=None, show=True):
-    """Plot sensors positions.
-
-    Parameters
-    ----------
-    info : Instance of Info
-        Info structure containing the channel locations.
-    kind : str
-        Whether to plot the sensors as 3d or as topomap. Available options
-        'topomap', '3d'. Defaults to 'topomap'.
-    picks : array-like of int | None
-        The indices of the channels to consider. If None, all the data channels
-        are plotted.
-    show : bool
-        Show figure if True. Defaults to True.
-
-    Returns
-    -------
-    fig : instance of matplotlib figure
-        Figure containing the sensor topography.
-
-    Notes
-    -----
-
-    .. versionadded:: 0.12.0
-
-    """
-    if kind not in ['topomap', '3d']:
-        raise ValueError("Kind must be 'topomap' or '3d'.")
-    if picks is None:
-        picks = _pick_data_channels(info, exclude=[])
-    if kind == 'topomap':
-        pos = _auto_topomap_coords(info, picks, True)
-    else:
-        pos = np.asarray([ch['loc'][:3] for ch in info['chs']])[picks]
-    def_colors = _handle_default('color')
-    ch_names = np.array(info['ch_names'])[picks]
-    bads = [idx for idx, name in enumerate(ch_names) if name in info['bads']]
-    colors = ['red' if i in bads else def_colors[channel_type(info, i)] for i
-              in picks]
-    return _plot_sensors(pos, colors, ch_names, show)
-
-
-def _plot_sensors(pos, colors, ch_names, show):
-    """Helper function for plotting sensors."""
-    import matplotlib.pyplot as plt
-    from mpl_toolkits.mplot3d import Axes3D
-    fig = plt.figure()
-
-    def onpick(event, pos):
-        ind = event.ind[0]  # Just take the first sensor.
-        ch_name = ch_names[ind]
-        pos = pos[ind]
-
-        # XXX: Bug in matplotlib won't allow setting the position of existing
-        # text item, so we create a new one.
-        ax.texts.pop(0)
-        if len(pos) == 3:
-            ax.text(pos[0], pos[1], pos[2], ch_name)
-        else:
-            ax.text(pos[0], pos[1], ch_name)
-        fig.canvas.draw()
-
-    if pos.shape[1] == 3:
-        ax = Axes3D(fig)
-        ax = fig.gca(projection='3d')
-        ax.text(0, 0, 0, '', zorder=1)
-        ax.scatter(pos[:, 0], pos[:, 1], pos[:, 2], picker=True, c=colors)
-    else:
-        ax = fig.add_subplot(111)
-        ax.text(0, 0, '', zorder=1)
-        ax.set_xticks([])
-        ax.set_yticks([])
-        fig.subplots_adjust(left=0, bottom=0, right=1, top=1, wspace=None,
-                            hspace=None)
-        pos, outlines = _check_outlines(pos, 'head')
-        outlines_ = dict([(k, v) for k, v in outlines.items() if k not in
-                          ['patch', 'autoshrink']])
-        for k, (x, y) in outlines_.items():
-            if 'mask' in k:
-                continue
-            ax.plot(x, y, color='k', linewidth=1, clip_on=False)
-        ax.scatter(pos[:, 0], pos[:, 1], picker=True, c=colors)
-
-    picker = partial(onpick, pos=pos)
-    fig.canvas.mpl_connect('pick_event', picker)
-    plt_show(show)
     return fig
